quick commit

+++ HI_FAST/test_light.py
@@ -102,7 +102,7 @@ def plot_coldgas(z):
     for i in range(len(model_names)):
         index = model_names[i]
         print model_labels[i],nGals[index],"objects"
-        pos =  gal[index]['Pos'].astype(numpy.float32)
+        pos =  gal[index]['Pos']
         c = numpy.empty((nGals[index],3),dtype=numpy.float32)
         # start = timer()
         # i = 0
